pyproject.toml
[tool.poetry]
name = "obsidian-parser"
version = "0.1.0"
description = "Script for parsing Obsidian markdown documents with frontmatter into a Pandas DataFrame"
authors = ["C4 Patino <c4patino@gmail.com>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
PyYAML = "^6.0.2"
pandas = "^2.2.3"
numpy = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
